# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Arrivo API Documentation",
        "description": "API endpoints for Savings, Groups, and User Management",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "localhost",
            "description": "Primary API Server"
        }
    ],
    "paths": {
        "/api/admin/users": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Admin: View all registered users",
                "operationId": "12d8068fb9094b67b1d7144ae2ee938d",
                "responses": {
                    "200": {
                        "description": "List of all users"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/users/{user}": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Admin: View details of a specific user",
                "operationId": "953c4b3c31f9587fc36549d06639865f",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User details"
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/groups": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Admin: View all saving groups",
                "operationId": "7a2c6d080e929170ab1bfc16a5ab47a3",
                "responses": {
                    "200": {
                        "description": "List of all groups"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/groups/{group}": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Admin: View specific group details",
                "operationId": "64560f6fa5840dc12071deeb41bb5053",
                "parameters": [
                    {
                        "name": "group",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Group details"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/saving-plans": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Admin: View all saving plans",
                "operationId": "2b0119428fd6c24f6a1265282e768416",
                "responses": {
                    "200": {
                        "description": "List of all plans"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/saving-plans/{plan}": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Admin: View specific saving plan details",
                "operationId": "5d850cbbb9a17d57e39cdccf746e0933",
                "parameters": [
                    {
                        "name": "plan",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Plan details"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/saving-plans": {
            "get": {
                "tags": [
                    "Savings"
                ],
                "summary": "List user saving plans",
                "operationId": "82d7a8f9fbca9b098bc2b1fbe57d78d6",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Savings"
                ],
                "summary": "Create a new saving plan",
                "description": "Creates a personal or group-based saving plan. If group_id is provided, the plan is owned by the group.",
                "operationId": "1a70addc80e2cb3f24080be54160f1e7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "description"
                                ],
                                "properties": {
                                    "description": {
                                        "type": "string",
                                        "example": "Save for New Laptop"
                                    },
                                    "target": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 1500,
                                        "nullable": true
                                    },
                                    "expire_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-12-31",
                                        "nullable": true
                                    },
                                    "group_id": {
                                        "type": "integer",
                                        "example": 1,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Plan created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}